Fixed Animation Bug

+++ source/extern/animation/Animatable.h
@@ -16,7 +16,7 @@ private:
     int startFrame;
     int endFrame;
     T value;
-    map<int, Keyframe<T> *> keys;
+    map<float, Keyframe<T> *> keys;
     
 public:
